Fix compilation for both platforms
