+++ lib/spindle/spindle.h
@@ -1,39 +1,2 @@
-#ifdef ESP32
-#include <ESP32Encoder.h>
-#else
-#include <Encoder.h>
-#endif
-#include <axis.h>
-#include <els_elapsedMillis.h>
-
-#pragma once
-
-class Spindle : public RotationalAxis {
- private:
-  // the unconsumed position is the position that has been read from the encoder
-  // but hasn't been used to update the current position of any driven axes
-  int m_unconsumedPosition;
-
-#ifndef ELS_SPINDLE_DRIVEN
-#ifdef ESP32
-  ESP32Encoder  m_encoder;
-#else
-  Encoder  m_encoder;
-#endif
-#endif
-
- public:
-#ifndef ELS_SPINDLE_DRIVEN
-  Spindle(int pinA, int pinB);
-#endif
-
-  void update();
-  void setCurrentPosition(int position);
-  void incrementCurrentPosition(int amount);
-  /**
-   * This will return the unconsumed position and reset it to 0
-   * used for updating the expected position of any driven axes
-   */
-  int consumePosition();
-  float getEstimatedVelocityInRPM();
-};
+#include "ESPSpindle.h"
+#include "TeensySpindle.h"
