+++ platformio.ini
@@ -25,7 +25,6 @@ lib_deps =
 	adafruit/Adafruit GFX Library@^1.11.9
 	adafruit/Adafruit SSD1306@^2.5.10
 	pfeerick/elapsedMillis@^1.0.6
-	lennarthennigs/ESP Telnet@^2.2.3
 
 [env:teensy41_debug]
 platform = teensy
@@ -39,7 +38,6 @@ lib_deps =
 	adafruit/Adafruit GFX Library@^1.11.9
 	adafruit/Adafruit SSD1306@^2.5.10
 	pfeerick/elapsedMillis@^1.0.6
-	lennarthennigs/ESP Telnet@^2.2.3
 
 [env:esp32dev]
 platform = espressif32
@@ -67,4 +65,7 @@ build_type = debug
 debug_test = *
 build_flags = -Wp,-w
 debug_build_flags = -O0 -g -ggdb
-lib_deps = lennarthennigs/ESP Telnet@^2.2.3
+lib_deps = 
+	lennarthennigs/ESP Telnet@^2.2.3
+	pfeerick/elapsedMillis@^1.0.6
+	google/googletest@^1.15.2
